[BUGFIX] Add read_only to cache identifier for file mounts
The cache identifier for file mounts only consisted of storage id
and path. In rare cases (with same name file mounts with different
read-only settings) this lead to rendering an arbitrary folder instead
of the mount with the "higher" access rights.

The read_only field is now taken into account when generating the
runtime cache identifier.

The file module already takes care of using the write file mount if
both read-only and not read-only exist.

Resolves: #99103
Releases: main, 12.4
Change-Id: If4c62b6e0dd8fd5a8e6e8581b072a4f22e27f075
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79654
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>