[BUGFIX] Do not show deleted page translation in page module in WS
This change does not render a page translation anymore
if it was deleted (t3ver_state=2).

It also enables the dropdown to create a new page translation
but then DataHandler does not allow this anymore.

This patch needs more work in DataHandler for
"delete translation in WS + create a new translation in WS"

Resolves: #98544
Releases: main, 12.4
Change-Id: Ie3a786f28faaf13c4165667ff8fc8a6d98f6e6db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79717
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>