[BUGFIX] Normalize filename of uploaded files
The filename of uploaded files might not be encoded
as normalized unicode. For instance, this happens when
using umlauts in filenames on HFS+ filesystem (macOS).

For instance the client sends an `ö`, which is sent in
NFD as `0x6fcc88`, but should be normalized as `0xc3b6`.

https://en.wikipedia.org/wiki/Unicode_equivalence#Normalization

Executed commands:
composer req symfony/polyfill-intl-normalizer:^1.27
composer req symfony/polyfill-intl-normalizer:^1.27 \
    -d typo3/sysext/core --no-update

Resolves: #101253
Releases: main, 12.4, 11.5
Change-Id: I8605481ffdc3b5d96f529850bf09a1fd75d09cd2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79838
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>