[BUGFIX] styleguide: Generate correct values for `datetime` type fields
The database generator does not use the DataHandler for
all operation due to some performance considerations.

That means, that special handlings included in DataHandler
needs to be adopted - at least in a stripped down version.

For example, the DataHandler contains value transformation
for `type = datetime` when the `dbType` is one of following
types: `datetime`, `time`, `timesec`.

Until now, now records are created with this constellation.
To be able to add such fields in the future, we need to mimic
partly work done by the DataHandler for these type in record
generation.

Releases: main, 12