[BUGFIX] Check default language for preview requirements
When previewing a translated page that is enabled which has
a parent (default lang) page that is hidden, users got a 403
error, because the preview flag was not set (as the page itself
was not hidden).

Change the check to always check both the default language and
the overlay for setting the preview flag.

Resolves: #20873
Resolves: #92775
Releases: main, 12.4
Change-Id: I438e772ee34cdb018cf4701878c18684a2fa3e03
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79813
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>