[BUGFIX] Do not store bollocks markup in icon cache
In very rare cases it may happen that a backend session is not valid
anymore when an icon is loaded within the TYPO3 backend. If the icon
is not in the user's localStorage, the TYPO3 backend login is stored
in localStorage instead.

To catch this issue in the future, additional checks were implemented:

* the response shall not come from a redirected request, which could be
  an indicator for unwanted markup
* the received markup must start with "<span"

Resolves: #101339
Releases: main, 12.4, 11.5
Change-Id: Ic30672b770b7d12aa078e423cb595966247d5506
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79985
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>