[BUGFIX] Make icon cache in localStorage version-aware
The identifier for the icon cache used in the client's localStorage is
now version-aware to get a chance of updated icons after a release. It
now uses the same identifier as the backend cache.

Previously, only the registry state itself was taken into account which
may not change if icons assets get updated.

Resolves: #101348
Releases: main, 12.4, 11.5
Change-Id: I15dfc9d4bf1288f9787f88f579beccf3b13a3fc6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80073
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>