[BUGFIX] Mark broken links in RTE again
The possibility to mark broken links with the `data-rte-error`
attribute and styling was removed in TYPO3 v12.

This is now added again to allow for linkvalidator and other
extensions to mark broken links which will then be visibly
styled differently when editing in the RTE.

This affects only the backend. The attribute data-rte-error is
removed when persisting the content to the database.

The logic for adding this attribute is still in place:
The linkvalidator uses the event `BrokenLinkAnalysisEvent`
which is dispatched in `RteHtmlParser`.

A marked broken link may look like this:

```html
<a href="t3://page?uid=42"
  data-rte-error="External link is broken">link</a>
```

Note that the functionality got disabled with the new ckeditor
in #96874 while the implementing plugin "showbrokenlinks" was
removed later with #98497. The CSS is therefore restored from
the state as of before #98497.

Resolves: #101357
Related: #96874
Related: #98497
Releases: main, 12.4
Change-Id: I898b1de1b9bb9bf334740dc3b7ec8eb70627bed4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80892
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>