[TASK] Allow f:link.action and f:uri.action without Extbase
This change uses the core LinkFactory (FE) for creating
links in f:link.action and f:uri.action when not executed
in an Extbase context in order to avoid booting up Extbase.

To operate in without Extbase context all Extbase related
arguments are required to be passed as there is no
fallback for determination based on Extbase configuration.

Note: This change uses the same boilerplate approach as #98474.
The intention is to replace that with an universally usable
Extbase URI builder in TYPO3 v13.

Resolves: #101729
Related: #100758
Related: #98474
Releases: main, 12.4
Change-Id: Ib5c27c46e9420947c0347c05c299cd67badb52ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80908
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>