[TASK] Drop redundant condition from TransportFactory
The variable `$port` in the changed code is an array element
from the return value of `GeneralUtility::trimExplode()` with
the `$removeEmptyValues` argument set to `true`. Hence it
is ensured to not be an empty string.

This change should be backported as it otherwise would block
improvements for the type annotations of
`GeneralUtility::trimExplode()`.

Resolves: #101384
Releases: main, 12.4, 11.5
Change-Id: I74ce433495fc3975c11991d21831db0244eacba0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80089
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>