[BUGFIX] Check if '_ORIG_uid' really exists in ElementInformationController
The computed property '_ORIG_uid' is only set for workspace records,
which are NOT moved (t3ver_state=4). It is better to check directly for
the existence of this key as done in all other occurrences where it is
retrieved. The presence already ensures that we are dealing with an
overlaid workspace record and that there is need to fetch the original
uid.

Resolves: #101426
Releases: main, 12.4, 11.5
Change-Id: I7dd1072736b476015d5d44cb82f5e670fdd484a7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80127
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>