[BUGFIX] Load importmaps for external form backend module scripts
The form migration to ES6 modules in #96568
missed to pre-register importmaps of third
party extensions.
The modules are loaded on initial request
via `import()`, but that requires an importmap
definition for external modules.

The modules are now preloaded in order to make
them available for dynamic imports.
Note that the modules are loaded by the form editor
and manager on initial load anyway, therefore it is
considered "ok" to preload the modules directly.

Resolves: #101373
Releases: main, 12.4
Change-Id: Ie7f5b05223729ffff24c7d76b211d977c9ed25dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80175
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan P <spo@snk.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan P <spo@snk.de>