[TASK] runTests.sh with podman, new CI
This patch is paired with the setup of a new CI [1].

The patch adapts runTests.sh to allow podman as
container environment next to docker. New CI relies
on podman.

Locally, podman can be selected as container environment
using the `-b podman` flag to runTests.sh. A "recent"
podman version - at least around 4.x - should be used.
docker is kept as default for now.

The new CI setup comes with a series of improvements
to speed up things. This makes the acceptance tests
a bit more brittle again. The patch comes with some
further stabilizations, and we may add even more with
dedicated patches later.

[1] https://git.typo3.org/typo3/CI/testing-infrastructure/

Change-Id: I206c1577335bfbefb801eaababa4b713e1edda12
Resolves: #102385
Releases: main, 12.4, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81744
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>