[TASK] Use `ignore-platform-req` workaround for `phpspec/prophecy`
The `phpspec/prophecy` defines a hard php version constraint
in the `composer.json` not allowing PHP8.3 for now. This may
tend to take quite some time and therefore the reason why it
has been removed in TYPO3 v12+.

To mitigate this issue way back for early PHP8.2 versions,
a workaround have been used to ignore the php platform req
only for the specific version. This should be reintroduced
for PHP8.3 now.

This change adds the `--ignore-platform-req=php+` workaround
again - but for PHP8.3 only.

Resolves: #101484
Releases: 11.5
Change-Id: I6b7eae083d35bc027d49811e645a8f83c648335e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80234
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>