[TASK] Remove superfluous check in extbase HashService
The `generateHmac()` function in extbase `HashService` checks
for the existence of the TYPO3 encryption key and throws an exception
if no encryption key is set. This check is however superfluous, since
`TYPO3\CMS\Core\Core\Bootstrap::checkEncryptionKey()` already does
the same.

The check in extbase `HashService` can therefore safely be removed.

Resolves: #101485
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Iae448a56ba738ac4f46e8c6b53c27525ffa99635
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80238
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>