[TASK] Map relative lit-imports to bare module specifiers
Imports via bare module specifiers (such that do not start with / or .)
are cache busted by their URL mapping in importmap — relative imports
(inside lit) are not by default. This is why we used version-suffixed
lit package folders in #96511 (pretty much like unpkg.com does).

This got removed with #100245, as required updates for configuration
files have not been applied automatically and the reason for the
version suffix was undocumented and thus unclear.
While we could fix those issues and add autoupdate support for
configuration files, it is now decided to remap all relative imports,
as this logic is present in our build setup anyway and eases future
updates by creating smaller diffsets (not every file has to be
renamed, only changed ones are updated).

Note that technically a new helper grunt task es-module-lexer-init
is added, which is required to await the asynchronous es-module-lexer
parser initialization (written in web assembly, and therefore
not synchronously available). This has worked "by luck" for the
copy:ts_files task (because there we earlier tasks that implicitly
awaited the initialization), copy:lit is forked off via the concurrent
plugin, we need to ensure es-module-lexer is actually ready to be used.

Resolves: #101504
Related: #101495
Related: #100245
Releases: main, 12.4
Change-Id: I7c714530d1a645da2d89179282f80bd730d12474
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80268
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>