[BUGFIX] Stabilize AbstractDomainObject _isDirty check
Accessing static properties as non-static has always
triggered a notice in PHP. However, since PHP 8.0
accessing undefined properties (which always comes with
the former notice) is now a warning. This will be
caught by the error handler in debug mode.

To fix this, the more robust _getProperty method is
backported from #100120 and used in the _isDirty method.

Resolves: #101488
Related: #100120
Releases: 11.5
Change-Id: I5f9ccb411f02f0dcc164d0d14917ecd631dc1178
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80280
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>