[TASK] Avoid reports.xml in acceptance tests
Creating reports.xml in acceptance tests and
feeding this to gitlab has been added with
issue #96628 in the hope to be useful for
finding frequently failing ac tests. The
gitlab UI however buries this, it's de facto
impossible to make use of this information.
We drop this definition again.

Resolves: #101514
Related: #96628
Releases: main, 12.4, 11.5
Change-Id: I91f13f402f0b9f72f2ca41252db93bb1714efcd6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80287
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>