[TASK] Fix MM TCA of blog_example
Extension blog_example is used in various extbase
tests. It defines invalid relations from tags
table over MM to post and persons.

The patch fixes the TCA - those relations now
work in BE as well, when the fixture blog_example
is activated.

Essentially all functional tests that works with
this extension break with the fixed TCA now. The
major part of the patch is to disentangle the .csv
import and export files to make them independent
from each other per test-case, and to fix the
imported and asserted db state along the way.

Resolves: #101528
Releases: main, 12.4
Change-Id: I4fb7d6ac82962f8ef7b630b261d543f4ff626bde
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80283
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>