[!!!][TASK] Remove `immediate` flag in `DebounceEvent`
With the introduction in TYPO3 v10, the `DebounceEvent` module had the
possibility to shift the event handler execution to the beginning of
the debounce sequence, enabled via the optional `immediate` parameter.

The parameter is unused in TYPO3 and using this feature has a negative
impact on UX. If used, the event handler is directly executed and the
user has to wait a specific time after the last event was triggered,
before any further execution is possible.

The flag `immediate` has been therefore removed.

Resolves: #101519
Releases: main
Change-Id: Ie0053fbfe3dcce4ae504550d024b1d88d69802f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80292
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>