[FEATURE] Introduce PHP attribute to autoconfigure event listeners
A new custom PHP attribute "AsEventListener" is added in
order to autoconfigure a class as an event listener.

The attributes is repeatable, which allows to register
the same class to listen for different events.

The attribute features the same functionality as
the manual registration in Services.(yaml|php) file.

Resolves: #101544
Releases: main
Change-Id: I81e3da5694f57ee0c7c3dd507d50ca79e2015dc1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80309
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>