[FEATURE] Improve event listener registration
It's now possible to register event listeners with
the `AsEventListener` PHP Attribute - introduced
with #101544 - on a method level.

Additionally, the listener identifier is now exposed
into the ListenerProvider, making this information
available, e.g. in the configuration module.

Registration of login failed related listeners is
adapted accordingly, which also fixes a duplicated
registration due to a missing `event` declaration.

Resolves: #101573
Related: #101544
Releases: main
Change-Id: I2ac228416b256513a919519692e74d7409df0d01
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80386
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>