[TASK] Mark PasswordPolicyValidator as public API
With the introduction of configurable password policies, the
`PasswordPolicyValidator` has been introduced. The class is
however marked as `@internal`, which basically was a mistake.
3rd party extensions providing users the possibility to change
a password should be able to validate the new password against
the globally configured password policy by using the
`isValidPassword()` method in the `PasswordPolicyValidator` class.

The change therefore mark the class as public API by removing
the `@internal` annotation.

Resolves: #101579
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I8869a8465199a2c071d3d780af640a0c903404c3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80357
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>