[BUGFIX] Don't notify the editor about canceled SvgTree Ajax request
If the svg-tree component is embedded in an iframe,
and if the iframe src get changed by navigating to another url,
pending AjaxRequest will be cancelled by the browser,
without letting us the opportunity to properly catch the thrown error.

As a workaround, we register an unload event on the iframe's window,
and mute notifications to the user.

Resolves: #101340
Releases: main, 12.4, 11.5
Change-Id: I8572fde4bf278481a32a459f802af69356ce28a5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80361
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>