[DOCS] Streamline the display of references to PHP class in EXT:form
In some cases, classes are displayed like "TYPO3CMSCoreTypeContextualFeedbackSeverity"
as backslashes are neither quoted not enclosed in a ":php:" directive.

Where appropriate, we now link to the source code directly instead of just mentioning a class (with ":t3src:)". This is helpful, for example,
for the severity enum, when you just want to know the available cases.

Additionally, other mentions of PHP classes are now enclosed in a
":php:" directive instead double ticks.

Resolves: #101599
Releases: main, 12.4
Change-Id: I2b384e388e6b6544c0a534a2bb23e97bc7c7fc1e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80418
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: core-ci <typo3@b13.com>