[TASK] Replace `php -S` with apache2 and php-fpm
This change uses a slighty modified httpd (apache2)
image based on alpine along with a own php-fpm image
to spawn up a webserver with php-fpm in two containers
as a replacement for the `php -S` webserver used prior.

Modification for the apache2 (httpd) image are minimal
and gives us the ability to mount the project with the
exact same path in the image as from outside. Further
all required apache2 modules are loaded missing the
official image.

Due to moving away from the php webserver we now need
to adjust some configurations:

* Use port 80 instead of 8000 the for acceptance
  and acceptanceInstall tests webserver.
* Use the acceptance test instance as DocumentRoot
  instead of the monorepo root.
* Copy several `.htaccess` files to the acceptance
  test instance as we now have a real apache2.
* Copy the favicon.ico to the test instance instead
  of the monorepo root.
* Ensure that created frontend SiteConfiguration no
  longer uses the relative test instance path.
  Simply use `/`.
* Use only the php-fpm image for cli and webserver
  fpm container.

Used command(s):

> composer u "typo3/testing-framework"

Resolves: #101827
Releases: main
Change-Id: I25367dd30963958c303bb91cea39adf4099f8718
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80568
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>