[TASK] De-mess EM InstallUtility a bit
Service classes in Extensionmanager\Utility
are a great example on "when class abstraction
goes wrong" - we're fighting with those for
lots of major versions already.

To tame this mess, this first patch applies a rather
straight streamlining to InstallUtility: Use constructor
property promotion, move public methods up in the class,
remove heavy-mocking and useless unit tests that
only check if certain sub-methods are called.

Resolves: #101643
Releases: main
Change-Id: If27bee98f22120ca739659c453c26fa1795c49d5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80499
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>