[BUGFIX] Unbreak FormEngine frontend due to "Translation behavior" bug
There's currently a bug in the DataHandler that kicks in when copying a
localized page that uses a "Custom" translation behavior on text
fields. Due to the bug, state is broken and the JavaScript part cannot
determine expected state, leading to a JavaScript error that leads to an
endless spinner.

This patch only fixes the JavaScript issue and renders a warning in the
console, but FormEngine is usable again.

Resolves: #101658
Releases: main, 12.4
Change-Id: I4fab1dae49928423c53906b0ae4fe5b4bddfaacd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80517
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>