[TASK] Always reset singleton instances for whole testcases
For (unit) tests that use singletons, the singletone instances should
always be reset for all tests from the corresponding testcase, not
only for individual tests.

This reduces the risk of new tests missing the singleton reset.
Also, it makes the code a bit shorter.

Also drop one instance of setting `$this->resetSingletonInstances`
where the property already was set to `true`.

Resolves: #101664
Releases: main
Change-Id: I5b2f962b108de0e36083a1a3593af20d813394dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80524
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>