[TASK] Propagate cache tags of cached page parts to page cache tags
When using the first-class TypoScript cache or stdWrap cache functions
the given cache tags of the content object are not correctly propagated
to pages on consecutive requests. The cache tags were not stored in the
cache entry, which is why they were never applied to the rendered page
cache once they were reused.

This patch adds support to store and restore the cache tag like #90775
did. Thus cache tags of subparts are now added as page cache tags.

[1] https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/Cache.html#cache-as-first-class-function

Resolves: #101663
Related: #90775
Releases: main, 12.4
Change-Id: Icb53c7708a4e1d4dca0e7eca754471756f4e2d63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80527
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>