[BUGFIX] Fix classloader dump after extension install
With #101596 event listener registrations were migrated to
PHP attributes. Service registrations may only be performed
on non-static methods with symfony/dependency-injection <6.3 [1].
This caused the 'non-composer-class-loader' events to be silently
skipped, which is why no class loading information was dumped.
ClassLoadingInformation event listeners are moved into a separate
service with an instance method to circumvent this bug in
symfony/dependency-injection.

Also stray and unneeded Services.yaml configuration is removed.
(public: false is the default and does not need to be set at all)

[1] https://github.com/symfony/symfony/pull/47066

Resolves: #101679
Resolves: #101681
Related: #101596
Releases: main
Change-Id: Ie18b8f1123364073a5df7a07b6f8bf71d27cd150
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80546
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>