[TASK] Deny global jQuery usage via eslint rules
This allows to drop a npm patch that had been applied
to the TypeScript @types/jquery before, in order
to achieve the same effect.
With eslint we can additionally provide a good
hint for the developer, what to use instead.

For the v12 backport @types/jqueryui is removed
additionally, as it conflicts with the bootstrap
jquery definitions and was unused. This showed
a small adaption that needed to be changed in
upgrade-docs, which has already been fixed
in main and is reapplied in here.

Commands executed:
  npm remove patch-package
  npm install
  npm remove @types/jqueryui

Resolves: #101697
Releases: main, 12.4
Change-Id: I77e15a4b4c2a18f8b72ec2a5f41c7e54a7966c30
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80686
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>