[BUGFIX] Allow a languageKey in non-Extbase f:translate VH
Even though the <f:translate> ViewHelper has a "languageKey"
argument, the parameter was only evaluated for Extbase
requests. However, it should also be possible for non-Extbase
Fluid parts - such as Fluid emails.

The patch adds the "languageKey" handling and refactors the
tests to rely on a fixture extension. Tests are added for
"languageKey" scenarios including fallback chains.

The new tests reveal a bug in extbase LocalizationUtility with
fallback chains along the way, keeps according tests as skipped,
and adds `@todo` for further investigation.

Resolves: #101710
Related: #98377
Releases: main, 12.4
Change-Id: I5251a69c385b50d844d8e24b363109e03596cc6d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80596
Tested-by: core-ci <typo3@b13.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: J�rg B�sche <typo3@joergboesche.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>