[BUGFIX] Avoid console error when tabbing through EXT:form inspector
…of a new form element, that has not yet been persisted.

Tabbing through an elemenet is considered an "input" by
the browser, which is why the "input" even is triggered.
The result is an empty input element (as before), but
an intent is made to delete the underlying property
from the form element object model.
When the object model is empty (remeber, it is a new element),
and an input field refers to a nested property
(e.g. properties.fluidAdditionalAttributes.placeholder)
removing an element (e.g. 'placeholder')
will fail when the parent has not yet been set
(e.g. properties.fluidAdditionalAttributes)

Resolves: #101750
Releases: main, 12.4, 11.5
Change-Id: I49a9ead343e671c3d0289bd6d969316b888d05eb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80733
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>