[BUGFIX] Fix GeneralUtility::validEmail() with spaces in email
`GeneralUtility::validEmail()` uses the package egulias/email-validator
for validating emails. It treats the email `test @testmail.com` as
valid, but triggers a warning that is not caught, since the syntax has
only been marked as deprecated by egulias/email-validator.

A check for warnings of the type CFWSNearAt will now be performed, which
results in the email `test @testmail.com` being marked as invalid.

Resolves: #101776
Releases: main, 12.4
Change-Id: I0808afaedfe5294b32dac68c0746fa725708f916
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82116
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>