[BUGFIX] Ensure correct record type for new record in SuggestWizard
FormEngine sends some information as payload with the suggest
wizard ajax request to the `SuggestWizardController` as context.
The controller loads the record based on the tableName and
the uid to determine the recordType to read the correct TCA field
configuration, respecting special overrides like `columnsOverrides`.

If a new record is created, there is no record uid available and
passed to the controller. Therefore, the recordType is not properly
determined. That leads to wrong record suggestions, if for example
the `allowed` record table is overriden for specific recordTypes.

This change adds the recordTypeValue as additional html data
attribute to the suggest search field, reads and sends it along
with the context payload in the ajax request to the suggest
wizard controller, which now uses the passed value while keeping
the record retrievement as fallback for now.

Suggest records are now directly searched correctly respecting
the full TCA configuration for the type, even for new records.

Additionally, uid is send as null instead of the string "NaN".

Used command(s):

> Build/Scripts/runTests.sh -s buildJavascript

Resolves: #101796
Releases: main, 12.4, 11.5
Change-Id: I3b814d37b7d4d3e9674ad6f2af882520c4f91413
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80773
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>