[TASK] Restore "ShowBlocks" functionality in CKEditor5
With the initial upgrade to CKEditor5 in TYPO3 v12, the 'ShowBlocks'
plugin was not available upstream, so this feature got lost.

As of CKEditor5 v38.1.0 [1] the plugin has been reintroduced.
We include the plugin now and add it to the Full RTE preset
configuration by default.

Executed commands:
  npm install --save @ckeditor/ckeditor5-show-blocks@38.1.0
  grunt build

[1] https://github.com/ckeditor/ckeditor5/releases/tag/v38.1.0

Resolves: #101330
Releases: main, 12.4
Change-Id: I82b4f3adbf70ed5beb101188d65a84c7e6b568f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80768
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>