[BUGFIX] Disable language (de)activation with write-protected configuration
The language management module writes into system/settings.php to
either enable or disable a language. With the changes made with #98957,
the language management should also inform about a write-protected
configuration file and remove any options that conflict with this
state.

This patch removes any option to either enable or disable a language and
moves the Fluid partial to a more generic location aimed for multiple
usages across several modules.

Resolves: #99008
Related: #98957
Releases: main, 12.4
Change-Id: Iae8c75e0727011289c421a66266989f44be76932
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80829
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>